# src/docs/openapi2/swagger.yaml
definitions:
  consts.ActionName:
    enum:
    - read
    - write
    - delete
    - execute
    - manage
    type: string
    x-enum-comments:
      ActionDelete: delete action
      ActionExecute: execute action
      ActionManage: manage action
      ActionRead: read action
      ActionWrite: write action
    x-enum-varnames:
    - ActionRead
    - ActionWrite
    - ActionDelete
    - ActionExecute
    - ActionManage
...
  consts.ResourceName:
    enum:
    - system
    - audit
    - container
    - container_version
    - dataset
    - dataset_version
    - project
    - label
    - user
    - role
    - permission
    - task
    - trace
    - injection
    - execution
    type: string
    x-enum-comments:
      ResourceAudit: audit resource
      ResourceContainer: container resource
      ResourceContainerVersion: container version resource
      ResourceDataset: dataset resource
      ResourceDatasetVersion: dataset version resource
      ResourceExecution: execution resource
      ResourceInjection: fault injection resource
      ResourceLabel: label resource
      ResourcePermission: permission resource
      ResourceProject: project resource
      ResourceRole: role resource
      ResourceSystem: system resource
      ResourceTask: task resource
      ResourceTrace: trace resource
      ResourceUser: user resource
    x-enum-varnames:
    - ResourceSystem
    - ResourceAudit
    - ResourceContainer
    - ResourceContainerVersion
    - ResourceDataset
    - ResourceDatasetVersion
    - ResourceProject
    - ResourceLabel
    - ResourceUser
    - ResourceRole
    - ResourcePermission
    - ResourceTask
    - ResourceTrace
    - ResourceInjection
    - ResourceExecution
...
  consts.StatusType:
    enum:
    - -1
    - 0
    - 1
    type: integer
    x-enum-varnames:
    - CommonDeleted
    - CommonDisabled
    - CommonEnabled
...
  dto.CreatePermissionReq:
    properties:
      action:
        $ref: '#/definitions/consts.ActionName'
      description:
        type: string
      display_name:
        type: string
      resource_id:
        minimum: 1
        type: integer
    required:
    - action
    - resource_id
    type: object
...
  dto.DateRange:
    properties:
      from:
        type: string
      to:
        type: string
    type: object
...
  dto.FilterOperator:
    enum:
    - eq
    - ne
    - gt
    - gte
    - lt
    - lte
    - like
    - starts
    - ends
    - nlike
    - in
    - nin
    - "null"
    - nnull
    - deq
    - dafter
    - dbefore
    - dbetween
    type: string
    x-enum-comments:
      OpDateAfter: DATE(field) > DATE(value)
      OpDateBefore: DATE(field) < DATE(value)
      OpDateBetween: DATE(field) BETWEEN date1 AND date2
      OpDateEqual: DATE(field) = DATE(value)
      OpEndsWith: LIKE %value
      OpEqual: ==
      OpGreater: '>'
      OpGreaterEq: '>='
      OpIn: IN (value1, value2, ...)
      OpIsNotNull: IS NOT NULL
      OpIsNull: IS NULL
      OpLess: <
      OpLessEq: <=
      OpLike: LIKE %value%
      OpNotEqual: '!='
      OpNotIn: NOT IN (value1, value2, ...)
      OpNotLike: NOT LIKE %value%
      OpStartsWith: LIKE value%
    x-enum-varnames:
    - OpEqual
    - OpNotEqual
    - OpGreater
    - OpGreaterEq
    - OpLess
    - OpLessEq
    - OpLike
    - OpStartsWith
    - OpEndsWith
    - OpNotLike
    - OpIn
    - OpNotIn
    - OpIsNull
    - OpIsNotNull
    - OpDateEqual
    - OpDateAfter
    - OpDateBefore
    - OpDateBetween
  dto.GenericResponse-any:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-array_dto_RoleResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        description: Generic type data
        items:
          $ref: '#/definitions/dto.RoleResp'
        type: array
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_PermissionDetailResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.PermissionDetailResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
  dto.GenericResponse-dto_PermissionResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.PermissionResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.GenericResponse-dto_SearchResp-dto_PermissionResp:
    properties:
      code:
        description: Status code
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/dto.SearchResp-dto_PermissionResp'
        description: Generic type data
      message:
        description: Response message
        type: string
      timestamp:
        description: Response generation time
        type: integer
    type: object
...
  dto.PaginationInfo:
    properties:
      page:
        example: 1
        type: integer
      size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  dto.PermissionDetailResp:
    properties:
      action:
        type: string
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      resource:
        $ref: '#/definitions/dto.ResourceResp'
      status:
        type: string
      updated_at:
        type: string
    type: object
  dto.PermissionResp:
    properties:
      action:
        type: string
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      resource_name:
        $ref: '#/definitions/consts.ResourceName'
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ResourceResp:
    properties:
      category:
        type: string
      display_name:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      type:
        type: string
    type: object
...
  dto.RoleResp:
    properties:
      display_name:
        type: string
      id:
        type: integer
      is_system:
        type: boolean
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.SearchFilter:
    properties:
      field:
        description: Field name
        type: string
      operator:
        allOf:
        - $ref: '#/definitions/dto.FilterOperator'
        description: Operator
      value:
        description: Value (can be string, number, boolean, etc.)
        type: string
      values:
        description: Multiple values (for IN operations etc.)
        items:
          type: string
        type: array
    required:
    - field
    - operator
    type: object
...
  dto.SearchPermissionReq:
    properties:
      actions:
        description: Action filter
        items:
          type: string
        type: array
      created_at:
        allOf:
        - $ref: '#/definitions/dto.DateRange'
        description: Common filters shortcuts
      description_pattern:
        description: Fuzzy match for description
        type: string
      display_name_pattern:
        description: Fuzzy match for display name
        type: string
      exclude_fields:
        items:
          type: string
        type: array
      filters:
        description: Filters
        items:
          $ref: '#/definitions/dto.SearchFilter'
        type: array
      include:
        description: Include related entities
        items:
          type: string
        type: array
      include_fields:
        description: Include/Exclude fields
        items:
          type: string
        type: array
      is_active:
        type: boolean
      is_system:
        description: Is system permission
        type: boolean
      keyword:
        description: Search keyword (for general text search)
        type: string
      name_pattern:
        description: Permission-specific filter shortcuts
        type: string
      page:
        description: Pagination
        minimum: 1
        type: integer
      project_id:
        type: integer
      resource_ids:
        description: Resource ID filter
        items:
          type: integer
        type: array
      resource_names:
        description: Resource name filter
        items:
          type: string
        type: array
      role_ids:
        description: Role IDs that have this permission
        items:
          type: integer
        type: array
      size:
        maximum: 1000
        minimum: 1
        type: integer
      sort:
        description: Sort
        items:
          $ref: '#/definitions/dto.SortOption'
        type: array
      status:
        items:
          type: integer
        type: array
      updated_at:
        $ref: '#/definitions/dto.DateRange'
      user_id:
        type: integer
    type: object
...
  dto.SearchResp-dto_PermissionResp:
    properties:
      applied_filters:
        items:
          $ref: '#/definitions/dto.SearchFilter'
        type: array
      applied_sort:
        items:
          $ref: '#/definitions/dto.SortOption'
        type: array
      items:
        items:
          $ref: '#/definitions/dto.PermissionResp'
        type: array
      pagination:
        $ref: '#/definitions/dto.PaginationInfo'
    type: object
...
  dto.SortDirection:
    enum:
    - asc
    - desc
    type: string
    x-enum-varnames:
    - SortASC
    - SortDESC
  dto.SortOption:
    properties:
      direction:
        allOf:
        - $ref: '#/definitions/dto.SortDirection'
        description: Sort direction
      field:
        description: Sort field
        type: string
    required:
    - direction
    - field
    type: object
...
  dto.UpdatePermissionReq:
    properties:
      action:
        $ref: '#/definitions/consts.ActionName'
      description:
        type: string
      display_name:
        type: string
      resource_id:
        type: integer
      status:
        $ref: '#/definitions/consts.StatusType'
    type: object
...
host: localhost:8082
info:
  contact:
    email: team@rcabench.com
    name: RCABench Team
  description: |-
    RCABench - A comprehensive root cause analysis benchmarking platform for microservices
    This API provides endpoints for managing datasets, algorithms, evaluations, and fault injections
    for root cause analysis in distributed systems and microservices architectures.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: RCABench API
  version: 1.0.1
paths:
...
  /api/v2/permissions:
    get:
      description: Get paginated list of permissions with optional filtering
      operationId: list_permissions
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      - description: Filter by action
        in: query
        name: action
        type: string
      - description: Filter by system permission
        in: query
        name: is_system
        type: boolean
      - description: Filter by status
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Permissions retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_PermissionResp'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List permissions
      tags:
      - Permissions
    post:
      consumes:
      - application/json
      description: Create a new permission with specified resource and action
      operationId: create_permission
      parameters:
      - description: Permission creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePermissionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Permission created successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_PermissionResp'
        "400":
          description: Invalid request format or parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "409":
          description: Permission already exists
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Create a new permission
      tags:
      - Permissions
  /api/v2/permissions/{id}:
    delete:
      description: Delete a permission (soft delete by setting status to -1)
      operationId: delete_permission
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Permission deleted successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "400":
          description: Invalid permission ID
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Cannot delete system permission
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Delete permission
      tags:
      - Permissions
    get:
      description: Get detailed information about a specific permission
      operationId: get_permission_by_id
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Permission retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_PermissionDetailResp'
        "400":
          description: Invalid permission ID
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Get permission by ID
      tags:
      - Permissions
    put:
      consumes:
      - application/json
      description: Update permission information (partial update supported)
      operationId: update_permission
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: integer
      - description: Permission update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePermissionReq'
      produces:
      - application/json
      responses:
        "202":
          description: Permission updated successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_PermissionResp'
        "400":
          description: Invalid permission ID or request format/parameters
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - Permissions
  /api/v2/permissions/{permission_id}/roles:
    get:
      description: Get list of roles assigned to a specific permission
      operationId: list_roles_with_permission
      parameters:
      - description: Permission ID
        in: path
        name: permission_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Roles retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-array_dto_RoleResp'
        "400":
          description: Invalid permission ID
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: List roles from permission
      tags:
      - Permissions
  /api/v2/permissions/search:
    post:
      consumes:
      - application/json
      description: Search permissions with complex filtering and sorting
      operationId: search_permissions
      parameters:
      - description: Permission search request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchPermissionReq'
      produces:
      - application/json
      responses:
        "200":
          description: Permissions retrieved successfully
          schema:
            $ref: '#/definitions/dto.GenericResponse-dto_SearchResp-dto_PermissionResp'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "403":
          description: Permission denied
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.GenericResponse-any'
      security:
      - BearerAuth: []
      summary: Search permissions
      tags:
      - Permissions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"